Fixing pipex cmd utils.

+++ 42cursus/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: sipyeon <sipyeon@student.42gyeongsan.kr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/16 03:03:46 by sipyeon           #+#    #+#             */
-/*   Updated: 2024/10/23 18:28:00 by sipyeon          ###   ########.fr       */
+/*   Updated: 2025/02/19 21:17:58 by sipyeon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		ft_p(void *arg, char *base);
 int		ft_d(int arg);
 int		ft_u(unsigned int arg);
 int		ft_x(unsigned int arg, char *base);
-int		ft_strlen(char *s);
+size_t	ft_strlen(char *s);
 void	ft_putchar(int c);
 void	ft_putstr(char *s);
 char	*ft_l_itoa(long n);
